requirements.txt
av==9.2.0
click==8.1.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
numpy==1.22.4
Pillow==9.1.1
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
PySide6==6.2.4
python-dotenv==0.20.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
shiboken6==6.2.4
tkVideo==0.1
tkvideoplayer==2.0.0
